function-search-v1
Search by function.A faster and more distinct searching method. Compared with traditional searching methods, this code provides a way to find the index directly through the target value, which greatly increases the searching speed and reduces memory usage. It can be nearly twice as fast as the original binary search when dealing with large datasets.